+++ confmap/converter/expandconverter/expand.go
@@ -16,51 +16,123 @@ package expandconverter // import "go.opentelemetry.io/collector/confmap/convert
 
 import (
 	"context"
+	"fmt"
 	"os"
+	"sort"
 
 	"go.opentelemetry.io/collector/confmap"
+	"go.opentelemetry.io/collector/internal/nonfatalerror"
+	"go.opentelemetry.io/collector/service/featuregate"
 )
 
-type converter struct{}
+var (
+	// raiseErrorOnUnknownEnvVarFeatureGateID controls whether to raise an error when an environment variable is used but not present.
+	raiseErrorOnUnknownEnvVarFeatureGateID = "confmap.expandconverter.RaiseErrorOnUnknownEnvVar"
+	raiseErrorOnUnknownEnvVarFeatureGate   = featuregate.Gate{
+		ID:          raiseErrorOnUnknownEnvVarFeatureGateID,
+		Description: "Raise an error when an environment variable is used but not present",
+		Enabled:     false,
+	}
+)
+
+func init() {
+	featuregate.GetRegistry().MustRegister(raiseErrorOnUnknownEnvVarFeatureGate)
+}
+
+type converter struct {
+	registry *featuregate.Registry
+}
+
+type stringSet map[string]struct{}
+
+func (s stringSet) Add(str string) { s[str] = struct{}{} }
+func (s stringSet) Merge(other stringSet) {
+	for str := range other {
+		s[str] = struct{}{}
+	}
+}
+func (s stringSet) Slice() []string {
+	var slice []string
+	for elem := range s {
+		slice = append(slice, elem)
+	}
+	// sort for reproducibility in unit tests.
+	// remove if ever in the hot path.
+	sort.Strings(slice)
+	return slice
+}
 
 // New returns a confmap.Converter, that expands all environment variables for a given confmap.Conf.
 //
 // Notice: This API is experimental.
 func New() confmap.Converter {
-	return converter{}
+	return newWithRegistry(featuregate.GetRegistry())
 }
 
-func (converter) Convert(_ context.Context, conf *confmap.Conf) error {
-	out := make(map[string]interface{})
-	for _, k := range conf.AllKeys() {
-		out[k] = expandStringValues(conf.Get(k))
+// newWithRegistry is useful for unit tests.
+func newWithRegistry(registry *featuregate.Registry) confmap.Converter {
+	return &converter{
+		registry: registry,
 	}
-	return conf.Merge(confmap.NewFromStringMap(out))
 }
 
-func expandStringValues(value interface{}) interface{} {
+func (c converter) Convert(_ context.Context, conf *confmap.Conf) error {
+	out := make(map[string]interface{})
+	unknownVars := make(stringSet)
+	for _, k := range conf.AllKeys() {
+		expanded, unknownExpanded := expandStringValues(conf.Get(k))
+		out[k] = expanded
+		unknownVars.Merge(unknownExpanded)
+	}
+
+	unknownVarsSlice := unknownVars.Slice()
+	var unknownErr error
+	if len(unknownVarsSlice) > 0 {
+		unknownErr = fmt.Errorf("failed to expand unknown environment variable(s): %v", unknownVarsSlice)
+		if !c.registry.IsEnabled(raiseErrorOnUnknownEnvVarFeatureGateID) {
+			unknownErr = nonfatalerror.New(fmt.Errorf(
+				"%w. Use %q to turn this into a fatal error",
+				unknownErr, raiseErrorOnUnknownEnvVarFeatureGateID,
+			))
+		}
+	}
+
+	if err := conf.Merge(confmap.NewFromStringMap(out)); err != nil {
+		return err
+	}
+	return unknownErr
+}
+
+func expandStringValues(value interface{}) (expanded interface{}, unknownVars stringSet) {
+	unknownVars = make(stringSet)
+
 	switch v := value.(type) {
 	case string:
 		return expandEnv(v)
 	case []interface{}:
 		nslice := make([]interface{}, 0, len(v))
 		for _, vint := range v {
-			nslice = append(nslice, expandStringValues(vint))
+			expandedVal, unknownExpanded := expandStringValues(vint)
+			nslice = append(nslice, expandedVal)
+			unknownVars.Merge(unknownExpanded)
 		}
-		return nslice
+		return nslice, unknownVars
 	case map[string]interface{}:
 		nmap := map[string]interface{}{}
 		for mk, mv := range v {
-			nmap[mk] = expandStringValues(mv)
+			expandedVal, unknownExpanded := expandStringValues(mv)
+			nmap[mk] = expandedVal
+			unknownVars.Merge(unknownExpanded)
 		}
-		return nmap
+		return nmap, unknownVars
 	default:
-		return v
+		return v, unknownVars
 	}
 }
 
-func expandEnv(s string) string {
-	return os.Expand(s, func(str string) string {
+func expandEnv(s string) (expanded string, unknownVars stringSet) {
+	unknownVars = make(stringSet)
+	expanded = os.Expand(s, func(str string) string {
 		// This allows escaping environment variable substitution via $$, e.g.
 		// - $FOO will be substituted with env var FOO
 		// - $$FOO will be replaced with $FOO
@@ -68,6 +140,14 @@ func expandEnv(s string) string {
 		if str == "$" {
 			return "$"
 		}
-		return os.Getenv(str)
+
+		// Use LookupEnv to check if environment variable exists
+		val, ok := os.LookupEnv(str)
+		if !ok {
+			unknownVars.Add(str)
+		}
+		return val
 	})
+
+	return expanded, unknownVars
 }